[Main] Key에 BitMask 적용
