+++ Server/MainServer/ServerDefine.h
@@ -21,18 +21,6 @@ namespace NETWORK_UTIL
 	//_NODISCARD const bool GetRecvOrSendPacket(const LPVOID);
 }
 
-namespace BIT_CONVERTER
-{
-	constexpr BYTE SEND_BYTE = (1 << 7);
-
-	/*_NODISCARD*/ BYTE MakeSendPacket(const BYTE inPacketType) noexcept;
-	/*_NODISCARD*/ bool GetRecvOrSend(const char inChar) noexcept;
-	
-	/*_NODISCARD*/ BYTE MakeByteFromLeftAndRightByte(const BYTE inLeftByte, const BYTE inRightByte) noexcept;
-	/*_NODISCARD*/ BYTE GetLeft4Bit(const BYTE inByte) noexcept;
-	/*_NODISCARD*/ BYTE GetRight4Bit(const BYTE inByte) noexcept;
-}
-
 namespace ERROR_HANDLING {
 	// 해당 static Function Array의 초기화는 GameServer의 생성자에서 이루어짐.
 	static std::function<void(void)> errorRecvOrSendArr[2];
